User can access entire access token response
As part of the Oauth2 login flow we get from the auth provider an access
token and potentially also a refresh token, id token, expiry date, and
token type. Of these `wai-middleware-auth` was only making the access
token available to the application behind the middleware, by way of the
`getAuthUser` function.

This commit gives access to the full response from the token provider
endpoint.

Not so nice is that the `getAuthUser` function now returns a
backwards-incompatible result for OAuth providers, without that being
explicit in the types. It might be nice to make an intentionally
breaking change to draw attention to this fact.