Google and Github pick access token from session
The session for an OAuth2 provider contains a number of tokens for those
that need them: an access token, id token, and refresh token might all
be available, depending on the settings of the OAuth2 provider and the
requested scopes.

The Google and Github OAuth2 providers only need the access token
though. This change picks it out of the client session for them.