Fix CI
By default the `clientsession` library obtains an encryption key to use
for client sessions from a file. When tests run in parallel they all try
to use this file at once, leading to test flakiness.

In this commit we override the default to use a generated in memory key
instead. This should not have the same concurrency downsides.